Tests/LibWeb/Text/expected/HTML/HTMLMediaElement-load-during-error.txt

Events:
  loadstart: readyState=0
  error: readyState=0
  error handler: setting good src: readyState=0
  error handler: src setter returned: readyState=0
  emptied: readyState=0
  loadstart: readyState=0
  durationchange: readyState=2 duration=12.043
  loadedmetadata: readyState=2
  loadeddata: readyState=2
  document load: readyState=2
  canplay: readyState=4
  canplaythrough: readyState=4
Final: readyState=4 duration=12.043
